Migrate connector builder server and UI to support APIs that return a non-object response (#118)
Migrates airbytehq/airbyte#20764 after the platform monorepo split

joe wrote the FE and I wrote the connector builder changes

For APIs with responses that were non-object responses encapsulated in { ... }. The connector builder read() would fail. It is perfectly reasonable for a API to response with an array of elements [ ... ] and potentially even string input. This change should allow for both.

The connector builder server no longer parsed response output into a json object because it might not be one. Instead we pass along the string output and let the frontend handle deserializing it back into the appropriate object to display back in the test output window