[Migration] [Connector Builder] Separate changes to incorporate Low Code Beta into Connector Builder UI (#102)
Migrated from airbytehq/airbyte#23129 to the airbyte-platform repo after the split.

We bumped the CDK_VERSION so the UI and builder server use the latest version of the language.

This includes FE changes that add additional transformations of the forms into YAML. There were also some changes to the forms themselves for concepts like stream slicer that were deprecated. The connector builder backend was modified to support the breaking changes to the language.