Add Workspace ID tag to Connection Manager's RecordMetrics activity (#4752)
Today we tag workflow metrics with the Connection ID in the RecordMetrics activity. I want to leverage these metrics to build a Canary monitoring dashboard, where we can track and monitor particular connections that we want to classify as "canary" connections. These are connections that we know to be stable and representative of a healthy platform, so any failures from these connections should trigger a monitor that engineers can investigate.

Ideally, we can create a Canary Workspace in which every connection is actively monitored for failures. This PR will make this easy by adding the WORKSPACE_ID as an additional tag/attribute in the RecordMetrics activity, so that the Canary monitor and dashboard can query for just the Canary's workspace ID, rather than needing to specify every Connection ID in the DataDog query.

- Fetch the connection's workspaceID using the workspaceApi in the RecordMetricsImpl.
- Wrap the API call in a try/catch so that a RetryableException can be thrown in case of any API connectivity issues.
- Update activity unit tests to cover new behavior (and some refactoring to DRY it up a bit)