fix(metrics): collect gauge values asynchronously
When using the "otel" collector for metrics-reporter.
You will get a log message like:

   WARNING: Instrument X has recorded multiple values for the same attributes.

This is because the Gauge builder in the OpenTelemetry Java SDK can only
collect gauge values asynchronously.

The old code essentially redefined the gauge every time with a callback
that produces the same value. The first registration succeeds since
there is no gauge defined; but subsequent gauge updates failed because
of the duplicate attributes.

This fix attempts to solve this by using a synchronous map and only
defining the gauge once. The callback uses the sync map to get the
latest value. Subsequent calls to gauge update the value in the map
rather than re-defining the gauge again.

This sort of a hack. The way that the API of OpenTelemetry is designed
is such that you should define your gauge up-front, passing a callback
which would perform the measurements periodically  by the SDK.
However, this API does not conform to the MetricClient interface
so this adapter is necessary unless the metrics client is refactored
to support this use-case.

Fixes #15623