fix(kubernetes): Exceptions in new KubePodProcess(...) can leak resources
If KubePodProcess fails to initialize its constructor the destroy()
method is never called, so the ports remain claimed. this will eventually lead
to port exhaustion.

This fix changes KubePodProcess constructor such that it catches any
Exception during the initialization phase.
If it does detect an error, it will call cleanup() which offers the
claimed ports back to the worker, and cleans up any other resources.
Finally, it will re-throw the exception to inform the caller.

Signed-off-by: Justen Walker <justen.walker+github@gmail.com>