fix(worker): ConnectionManagerWorkflowImpl null connectionId
An invariant of the ConnectionManagerWorkflow objects is violated when
the class is constructed by the runtime; namely the connectionId value is
null when it is expected to always be non-null.

This change detects a null connectionId in ConnectionManagerWorkflowImpl::run
and will set it to connectionUpdaterInput.getConnectionId() which is guaranteed
not to be null.

Signed-off-by: Justen Walker <justen.walker+github@gmail.com>