Create a form to clone the repo inside the pod k8s
The purpose of this modifications is to "re-clone" the repo inside the pod
if the folder doesn't exists. It reuses the structure of the NormalizationRunner.

At this points it uses a custom script to create the profile of bigquery destination.
The better form is to reuse the transform-config.py script to create the dbt profile.