Destination S3: Support IRSA roles
This change introduces support for using [IAM Roles for Service
Accounts (IRSA)](https://docs.aws.amazon.com/emr/latest/EMR-on-EKS-DevelopmentGuide/setting-up-enable-IAM.html)
within S3 destinations in Airbyte. It is based on the changes
in #177, but does not attempt to add the same support for logging.
The change should also work for S3 sources, but is not tested
thoroughly with sources, only with destinations.