move connection pausing logic into cloud-cron
## What

The bootloader is currently running sync pausing logic. This requires us to correctly determine any connector version overrides so we don't disable syncs for pinned folks (like PbA users), as we found in https://github.com/airbytehq/airbyte-platform-internal/pull/8889. We can't use the same `CloudDefinitionVersionOverrideProvider` in the bootloader since it uses the cloud-server and when the bootloader runs there is no cloud-server to call. 

We don't really need to pause syncs in the bootloader and it's a cloud only thing anyway, so this moves the sync-pausing logic out of SupportStateUpdater and into a new recurring task in our new `cloud-cron` app. 

It also sets the frequency to every 30 minutes since we don't really need to be doing this every 30 seconds (which is how often the current cron runs, and is what we want to get connector updates quickly available)

## How

- Create new ConnectionSupportStateValidatorCron class
- Move SupportStateUpdater code relevant to sync disabling into this new cron task


## Can this PR be safely reverted / rolled back?
*If you know that your PR is backwards-compatible and can be simply reverted or rolled back, check the YES box.*

*Otherwise if your PR has a breaking change, like a database migration for example, check the NO box.*

*If unsure, leave it blank.*
- [x] YES 💚
- [ ] NO ❌