Run the orchestrator in local kube (#9469)
Start an orchestrator pod when deployed in local kube. In order for that to work, we need to ensure that the bucket exist in the local doc store.

In the process, we discovered that the orchestrator was not turned on correctly for all OSS Helm charts.

We also realised some of the tests were not entirely correct. This PR attempts to fix all of this.

Co-authored-by: Davin Chia <davinchia@gmail.com>