Mirror Keycloak. (#9745)
Keycloak is hosted on Quay, which is more unreliable than reliable. Their status page shows on average 1 incident a for the last 3 months. Each time this takes down all builds.

Instead of always reaching out to quay for the Keycloak image, we mirror the image under our own airbyte repository.

This shields us from Quay downtimes and allows us to use the pull-through cache, which will speed up builds, especially since we do not change this image often.

Mirror requires a multi-arch image, which was slightly more complex than I thought. However I was able to boil it down into 5 steps, which I've put in a readme whenver we upgrade this image in the future.

This feels like acceptable complexity for the reliability tradeoff, given we don't upgrade Keycloak often.