Prefix platform logs with 'worker' for better UI display. (#10011)
There was a Q4B ask to better support the UI filtering logs by origin.

After chatting with @josephkmh, we realised the UI doing its own filtering isn't too bad since the UI already has to go through all logs to convert the ascii color -> HTML tag. The main thing blocking this for now is the untagged lines, and this is solvable as these log lines are emitted by the workers.

This PR adds the worker prefix to all these lines. I'll add screenshots in the comments to show this works.

- Inject this at the Temporal Attempt Execution layer, which wraps all UI-visible activity.
- Inject at the append log activity - afaict this is the only UI-visible workflow log we have.

With the exception of stack traces which are emitted with line breaks, and exploded out by the UI, all log lines should have origin prefixes, in both Docker and Kube.