add data service for scoped configuration (#10227)
## What

Adds and implements a new service to manage scoped_configuration. 

close https://github.com/airbytehq/airbyte/issues/32940

## How

In general  we're following the new pattern of using specific services by introducing a ScopedConfigurationService, but this uses micronaut-data rather than jooq since we want to move in that direction anyway. It also does so in kotlin.

External users would interact with the data using ScopedConfigurationService, which internally executes queries using the ScopedConfigurationRepository. The service exposes and works with config models defined in YAML/JSONSchema as is customary, and the micronaut data repository uses its own data classes.

- Add new ScopedConfigurationService to airbyte-data
- Create an implementation of the service using micronaut data
    - Add any dependencies / things needed to work with micronaut-data and kotlin in airbyte-data
    - Add a ScopedConfigurationRepository and relevant ScopedConfiguration entity
    - Add config-models for ScopedConfigurationa and related types
    - Add mappers to go between entities for micronaut data and generated config models

## Recommended reading order
1. `entities/ScopedConfiguration.kt`
2. `ScopedConfigurationRepository.kt`
3. `config-models/**`
4. `ScopedConfigurationMapper.kt`
5. `ScopedConfigurationService.kt`
6. `ScopedConfigurationServiceDataImpl.kt`
7. tests

## Can this PR be safely reverted / rolled back?

- [x] YES 💚
- [ ] NO ❌