fix apply definitions helper test behavior (#10688)
## What
* Weird behavior! Fix an incorrect test setup causing a sort-of false positive

The following behavior didn't stand out to me until I tried to add the docker repositories into the assertions. 

_What I intended_:
To set up a registry with 3 different definitions each, with an invalid definition in the middle, to make sure that both the first and last definition still got processed, expecting 4 INITIAL_VERSION_ADDED events

_What I did_: 
Accidentally used 2 different versions of the same source as the first and last entries in the registry

_What I would have intuitively expected to happen_: 
Instead of 4 INITIAL_VERSION_ADDED events, I was expecting an INITIAL VERSION_ADDED and a DEFAULT_VERSION_UPDATED for each

_What actually happens_: 
Because we get the list of Ids of connectors in our db at the very start, both versions are calculated as "net new" connectors. Since we upsert instead of splitting up insert/update into 2 actions with the relevant restraints, our platform doesn't actually have an issue trying to insert a new connector twice - only the metrics (and the counters for the logs I guess) actually distinguish it

_Why I think this is fine_: 
Pretty sure we can't create a registry with multiple entries for a given actor definition ID, so this was just me accidentally creating an invalid registry in the test. I don't think it's worth checking for an updated list of entries in the db before every definition. IMO the real way to fix this edge case that is not likely to ever happen in production is to have actual create/update operations (even micronaut data doesn't have a native upsert! :D) 

## How 
* Create a distinct 3rd definition in the list for testing

## Can this PR be safely reverted / rolled back?
*If you know that your PR is backwards-compatible and can be simply reverted or rolled back, check the YES box.*

*Otherwise if your PR has a breaking change, like a database migration for example, check the NO box.*

*If unsure, leave it blank.*
- [x] YES 💚
- [ ] NO ❌