Separate Worker Base Image. (#10975)
As I was iterating on builds over the last week I ran into several errors that happened due to timeouts while downloading binaries for install.

This PR solves that by creating a new worker base image that contains docker and kubectl. This prevents the need to repeated install these binaries on build.

The main goal is to help with flakiness. There should also be minor (1 or 2 min) speed improvements.