remove breakingChangesHelper method duplication (#10978)
## What
Found out a method that I am changing upstack is actually duplicated. Presumably this happened in the move to `airbyte-data` because they didn't want to add a dependency on config-persistence. I noticed because the old method had tests and the new one doesn't. 

This PR removes the old method and moves the tests to a new file. Note that the existing code in the `ConnectorMetadataJooqHelper` is already covered by various `XPersistenceTests`.

Once we are able to fix the transactionalization of this behavior, we can move this logic into helpers somewhere further away from airbyte-data. 

## Can this PR be safely reverted / rolled back?
*If you know that your PR is backwards-compatible and can be simply reverted or rolled back, check the YES box.*

*Otherwise if your PR has a breaking change, like a database migration for example, check the NO box.*

*If unsure, leave it blank.*
- [x] YES 💚
- [ ] NO ❌