ScopedConfigurationService: list by origin, batched create/delete/resolve (#11440)
## What

When dealing with breaking change pins, we'll need to potentially manage lots of scoped configurations at once. This PR adds some bulk methods to make things easier and more performant in these cases.

Specifically, four new methods are exposed as part of the ScopedConfigurationService:
- Create configurations in bulk
- Delete multiple configurations by IDs
- List configurations by origin type + values
- Resolve configuration values w/ scope map in bulk

The first three are pretty straightforward- the last one is where I'd want to get some closer eyes on.

## How

- Add necessary repository methods 
- Implement new data service methods using aforementioned repository methods

The notable change is in how we resolve configs for a scope map in bulk. We now grab all configs whose scopes are in the map, and resolve the right config based on the Key order as before.

The input map here is in the following form:
```json
{
    "key_1": {
        "actor": "actor_id",
        "workspace": "workspace_id"
        "organization": "organization_id"
    },
    "key_2": {
        "actor": "actor2_id",
        "workspace": "workspace_id"
        "organization": "organization_id"
    }
}
``` 

This is a map with unique keys rather than just a list since the caller needs to be able to identify the result for each set of scopes. As such, the output looks like:

```json
{
    "key_1": "<Resolved ScopedConfiguration object>,"
    "key_2": "<Resolved ScopedConfiguration object>",
}
```

If a key doesn't have a configuration, it's excluded from the map.

Usually these keys will be an ID of whatever entity you're resolving the config value for. E.g. an actor ID.

## Recommended reading order
1. `ScopedConfigurationService.kt`
2. `ScopedConfigurationRepository.kt`
3. `ScopedConfigurationServiceDataImpl.kt`
4. Tests

## Can this PR be safely reverted / rolled back?
*If you know that your PR is backwards-compatible and can be simply reverted or rolled back, check the YES box.*

*Otherwise if your PR has a breaking change, like a database migration for example, check the NO box.*

*If unsure, leave it blank.*
- [x] YES 💚
- [ ] NO ❌