Heartbeat: Use Default Timeout if configured timeout is too low. (#11875)
Part of the No Stuck Syncs project is an audit for API Sources - all Sources will now specify their maxSecondsBetweenMessages field according to their rate limits.

Certain sources have very low rate limits. We still want the field specified, so we have the data, however we are going to default to the default rate limit if the given rate limit is too low to start to minimise any production impact.

If the configured field is lower than the default, use the default.