Add the ability to expire ephemeral secrets. (#11961)
A recent increase in Cloud costs exposes how inefficient we can be with Secrets.

We never delete the secrets created for Check and Discover connections. This resulted in ~120k unused secrets that was costing us about 7k/month. Let's stop being dumb!

I've manually deleted these secrets. This PR is to make sure we stop doing this going forward.

For ephemeral secrets, we set an expiry time of 2 hours. This allows GSM to expire the secrets after.
This is currently only implemented for GSM, and defaults to the previous write command for other implementations.

I also added a script I was using to delete secrets. If people don't think it's useful, I can remove it from this PR.