Deduplicate default and realtime secret persistence methods. (#12284)
The endgoal here is to stop secrets from eating too much into our costs. We spend ~1k/day now on secrets!

To do that, we want to start deleting secrets when an actor's config is updated, and when an actor itself is tombstoned.

I have taken several stabs at this. I finally conclude the secret interface is quite convoluted today and step one is to clean up the SecretRepositoryWriter interface to be more general purpose - deep and short rather than shallow and wide.

First pass here is to deduplicate the runtime and default secret persistence methods.

Effect:
- less code path - easier to figure out the right place to insert deletion logic.
- eventually we want this to be backed by a database table, and not configs embedded in secrets. This puts the general interface closer to CRUD, making it easier for that transition to happen when we do want to make it happen.

Change:
- SecretsRepositoryWriter.kt - this is the main interface change.
-All other files are client side changes.