Add transforms for source-defined PK changes (#12312)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->

Part of https://github.com/airbytehq/airbyte-internal-issues/issues/7494

We want to consider source-defined PK changes as transforms in order to show them in the UI and properly handle breaking changes.

This adds a new concept of a `StreamAttributeTransform` alongside the existing `FieldTransform`s as part of a stream update. This was structured as a separate type from the field transforms since it doesn't quite represent a change in a field, but rather a change to the stream itself.  

Currently there is only one type of StreamAttributeTransform, which represents a primary key update. However, there could be more in the future, for example for handling source-defined cursor changes or anything else about the stream that doesn't have to do with fields.


## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌