Move from Postgres 13 -> 15. (#12341)
We recently upgraded to Postgres 15 internally. This PR moves our tests to 15 as well. The goal here is to be start getting ready to publicly say we support 15.

We want to do this because:
- 13 is 4 years old now and EOL is a year or so.
- 15 is something enterprise users are actually asking us for. It behooves us to state a more up-to-date infrastructure dependency now while the number of enterprise users are small so we avoid supporting a larger swath of database versions later on.
- lastly, staying up-to-date with software upgrades is best practice.

Bump all the tests to 15 to start enforcing this throughout the code base.

This PR does not modify our external docs or our docker/helm deployments. This is something Deployments will take on in the near future.