Differentiate between Secrets Create and Update (#12360)
Follow up to #12357 .

Simplify the SecretRepositoryWriter interface in preparation for deleting code.

After this refactor, adding the code for deleting dangling secrets when updating secrets is just a few lines of code change. Will do in a follow up PR to keep things simple.

## How
Instead of always calling update, call create when creating and update when updating.

This allows us to simplify the update method. Optionals go away. Always validate the spec.

At the same time, simplify the statefulSplitSecrets method to take in JsonNode and not the parent object for better scoping. This means all SecretRepositoryWriter methods now speak JsonNode. Eventually we want everything to be typed, however that's for a different project.