fix DefaultjobPersistenceTest (#12353)
`DefaultJobPersistenceTest` is consistently failing on a localhost based in CA (might be passing if your laptop is in another timezone)
Issue was tracked down to the timezone used to write attempts vs the timezone used to read them
When writing an attempt through`createAttempt`, `failAttempt` and `succeedAttempt`, and on reads, we were using `SYSTEM_OFFSET` that's set at `ZonedDateTime.now().getOffset()`.
Part of the problem is that when converting a date like `2021-01-01T00:00:00Z`, we get a different result when using the `ZoneOffset` vs using the `ZoneId`. And so if we use one to read, and another to write, we might not get the data we expect.
This PR fixes the test by always using `SYSTEM_TIMEZONE` in `DefaultJobPersistence`