Refactor: Simplify Secrets Update (#12497)
Following #12478, we learnt that if the passed in secret value doesn't change, we don't increment the coordinate version, and instead write the same value to the same coordinate.

This is not only confusing, it makes implementing deletes tricky. We now have to communicate whether a coordinate's version was updated or not up the call chain. This is not straightforward.

Instead, we keep things simple, and always increment the coordinate version regardless of secret value. This simplifies the call chain and is easy to reason with. Though we incur an additional secret, we are going to implement the deletion logic relatively soon anyway.