refactor: replace cdkVersionProvider with an injected bean (#12531)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->
The CDK version provider is now only used to get the CDK version used in the server so that we can return it for health checks. Simplifies how we retrieve and store this version.

## How
<!--
* Describe how code changes achieve the solution.
-->
Instead of creating a provider and a cached provider, load the resource within the ApplicationBeanFactory and inject the version into the health service that uses it. This caches it by default. 

Confirmed that the health check still has access to this resource by calling the health check endpoint from another pod: 
```
curl airbyte-connector-builder-server:8080/v1/health
{"available":true,"cdk_version":"0.89.0"}
```

## Recommended reading order
0. `oss/airbyte-config/init/src/main/kotlin/io/airbyte/config/init/DeclarativeSourceUpdater.kt`
1. `oss/airbyte-connector-builder-server/src/main/java/io/airbyte/connector_builder/config/ApplicationBeanFactory.java`
2. `oss/airbyte-connector-builder-server/src/main/java/io/airbyte/connector_builder/handlers/HealthHandler.java`
3. Others - delete code based on simplification

## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌