feat(db): add declarative_manifest_image_version table to support builder source auto-updating (#12486)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->
Adds a new db table, `declarative_manifest_image_version`, to keep track of the latest version of `airbyte/source-declarative-manifest` for each major version. This allows us to keep users as up-to-date as possible with cdk changes, without shipping breaking changes to them.

Of note: the versioning for connector builder projects and `source-declarative-manifest` is very different from standard connectors. We lean into this here - this is vaguely similar as the introduction of the `actor_definition_version` table. The long-term vision is to stop hacking custom specs into `airbyte/source-declarative-manifest` actor definition versions, so that source-declarative-manifest can have the standard one-entry-per-version in the `actor_definition_version` table like other sources, which can be shared by connector builder connectors (including the low-code connectors published from our repo!). But we're not quite there yet :) 

## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [ ] YES 💚
- [x] NO ❌