chore: move public api problems to new types (#12500)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->

Migrates our current (Legacy)AbstractThrowableProblems to the new generated types by adding them to our openapi spec and swapping the usages to the new classes.

## How
<!--
* Describe how code changes achieve the solution.
-->

The approach taken here is not to make new error types, but rather keep the existing behavior that the Public API had, just on the new generated problems system. This avoids confusion with having two ways of defining problems while letting us incrementally improve the errors as needed.

For example, some problems that are generic UnexpectedProblem or BadRequest just have a message with the details of the problem as they were before, but they should really be their own problems with more specific types.

The only change API users should see after this is merged is:
- `message` field becomes `detail`. This was actually already what the docs said it _should_ be, but in practice it was coming through as `message`.
    - Details are more standardized. Specific messages for generic problems like `UnexpectedProblem` are contained within `data.message`.

I've maintained the same `type` and `title` fields for existing problems. 

For reviewers, all changes should be straightforward replacements of the previous class with the new generated problem class. There should be no logic changes in here.

## Recommended reading order
1. `api-problems.yaml`

## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌