bug: ignore errors when disabling a secret (#12612)
Not doing so interrupts saving the configuration, leading to out-of-date database configuration that points to a disabled secret. This leads to user errors.

Ignore any disable errors we might run into.

This is fine since any errors is an artifact of legacy secret versioning. In the new world, we always create a new secret, so every secret only has one version. Disabling multiple versions no longer applies as we move forward.

Further, we are going to switch to deleting secrets in a week or so, so this is a short term fix.