feat: increase max successive attempts to 20. (#12709)
See this slack thread for detail: https://airbytehq-team.slack.com/archives/C03AS1GAQV6/p1717629461674829

TLDR: increase the number of partial failure attempts to account for the new CDC algorithmn. This also increases the chance that a job actually finishes, which is a good thing.

The trade off is longer jobs, however since we only reattempt if records move, this is a good thing.

I'm adjusting the default instead of using the feature flag because I want to keep OSS/Cloud in sync.