chore: upgrade build cache plugin to latest. (#12970)
As part of airbytehq/airbyte-platform#337 I was trying to figure out how to why we have the 1.x and 2.x AWS SDK dependencies in our deps.toml today. Either one should be sufficient.

As part of this I noticed our build cache dependencies are out of date so I'm going to update this.

Upgrade this to the latest version - https://github.com/burrunan/gradle-s3-build-cache.