feat: prepare worker v2 release (#12875)
Disable the workload on enterprise by default
Change the activation logic for using the workload. Instead of the old feature flag, we have added 2 new feature flags based on the env variable. Those variable are defined in the env-configmap and added to the worker.

In order to activate the workload, the enterprise customer have to modify the deploy.sh scripts and switch the enabled variable to true.

For backwards compatibility, we keep the existing Cloud Worker V2 feature flags. This lets us simplify the OSS configuration while leaving the Cloud configuration untouched.

We'll be able to rip everything out after we are done rolling this out.

Co-authored-by: Davin Chia <davinchia@gmail.com>