fix: add image pull secret to authenticate to docker. (#13070)
We had an incident today where we realised the orchestrator does not use docker secrets to pull images. This led to us being rate limited on pulls, causing issues in the EU cluster.

Use the image pull secret so we appear as authenticated and have a higher rate limit.

This is a hot fix. We are currently using the discover secret. we should follow up and plumb the correct configs in.