feat: add DropdownButton component (#13231)
## What
This is a pre-requisite for https://github.com/airbytehq/airbyte-internal-issues/issues/8337. As you can see in the [designs](https://www.figma.com/design/dT3VJHoe9joDnSkku4fh77/Publish-to-Github-from-Connector-Builder?node-id=4004-41&t=xS6Sg81NkRhtgk7T-0) for that issue, we need to render a button which has a dropdown menu attached to it.

This PR adds this in the form of the DropdownButton component.

## How
I started out by updating the Button component to support a `dropdown` prop, but ended up deciding against this approach because:
- It required several conditionals for which top-level container to apply button styles to which got messy fast
- This made the change more risky as it would have touched the Button component which is used all over the place

Instead, I opted to add a new DropdownButton component which contains two Buttons inside of it — one for the main button, and one for the dropdown caret. 

The benefits of doing it this way are:
- The existing button styles we want are applied to each of these two inner buttons individually, which is desired (e.g. hovering over the main button only changes the color of the main button, and hovering over the dropdown caret only changes the color of the dropdown caret button)
- The Button component can remain more simple, which is important for such a widely-used component.

I would recommend testing this by running `pnpm storybook` in the airbyte-webapp folder, and opening the `DropdownButton` UI component:

![Screenshot 2024-07-23 at 6.23.13 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/qHuCzMeboOsCv5wz21VS/2b50b033-c902-4fc9-adea-ea9c02093cbf.png)

## Recommended reading order
1. `oss/airbyte-webapp/src/components/ui/DropdownButton/DropdownButton.tsx`
2. `oss/airbyte-webapp/src/components/ui/DropdownButton/DropdownButton.module.scss`
3. `oss/airbyte-webapp/src/components/ui/DropdownButton/DropdownButton.stories.tsx`
4. `oss/airbyte-webapp/src/components/ui/DropdownMenu/types.ts`
5. `oss/airbyte-webapp/src/components/ui/DropdownMenu/DropdownMenu.module.scss`

## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌