feat: enable Workloads for OSS, continue opt-in for Enterprise. (#13362)
Enable Workload by default for OSS, continue opt-in for Enterprise.

While we are confident enough to release this to OSS, we still want this to be opt-in for Enterprise just in case. After a 2-week bake-in period, we'll only keep Workloads.

Enabling Workload means:
Deploying both the workload api server
Deploying the launcher
Public docs have already been updated for this change.

Introduce a migration workloadsOptIn flag under the enterprise block.

If the edition is community, deploy the workload api and launcher and set the right flags.
If the edition is enterprise and workloadsOptIn is true, then deploy the workload api and launcher and set the right flags. The assumption being enterprise users always set edition to enterprise.

Both services are always on, and we switch on these logic statements, instead of whether the chart is enabled or not.

Note: I tried to use helpers and run into enough issues with syntax and debugging the call chain I switched to simple if conditions since this is temporary.

Co-authored-by: Bryce Groff <bgroff@hawaii.edu>