chore: add tracking for contribute form focus (#13912)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->
Add a tracking event when the user focuses the Description field of the contribute to airbyte form. 

Our current fullstory funnel only looks at whether the user clicked the contribute radio button or selected it in the dropdown. This doesn't give us a great signal that they were actually trying to contribute, because a lot of people see a new thing to click and click it just to see what it's about, and then click away. By the time the user has clicked on description, the first field that is not auto-filled, we assume they are probably actually interested in starting a contribution. This step comes before the filling in the Github token step, which we assume may be a drop off, so getting this signal before the form submit is important. 

## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌