fix: fix build from incorrect regex due to uncontrolled dependency change. (#14708)
Something changed in our dependencies that caused this to fail. This required cleaning out the local M2 and gradle caches to pass. I never did track down what precise dependency caused this since my priority is to get this working to unblock everyone.

Update the syntax to the right text.