fix: better default temporal settings. (#14821)
Closes airbytehq/airbyte#30691 (comment).

This implements the fix suggested in airbytehq/airbyte#30691 (comment).

We are also seeing this in non-Docker deployments, regardless of our Worker refactors. This confirms this is needed to stablise OSS deployments with medium - high load. Temporal tells us this is fine as long as the OSS deployment has enough resources.