feat: add definitionId to enterprise stubs (#15224)
## What
<!--
* Describe what the change is solving.
* It helps to add screenshots if it affects the frontend.
-->

Adds `definitionId` to enterprise source stubs, and a new endpoint to fetch stubs for a workspace. This endpoint will filter out any definitions that have already been licensed to the given workspace/org.

Also refactors things a bit to make it more resilient by being able to deserialize entries with unknown properties, as well as returning an empty list rather than failing when something goes wrong when fetching the stubs.

Note that this doesn't mean we can add definitionId to the stubs yet since it doesn't address OSS folks on older versions. To be able to do this safely, we'll need to publish the stubs to a new URL and update the configured urls accordingly.


## Can this PR be safely reverted and rolled back?
<!--
* If you know that your be safely rolled back, check YES.*
* If that is not the case (e.g. a database migration), check NO.
* If unsure, leave it blank.*
-->
- [x] YES 💚
- [ ] NO ❌