chore: log out context to debug flag. (#15689)
This feature flag is not activating for some reason. This log line seems to show the correct context locally as well as in a dev env.

I'd like to see this line in prod just in case there is something special about prod that is causing this to not work. It's somewhat icky.. but it's the most surefire way of understanding what's happening.

Open to pushback if folks have suggestions.

Co-authored-by: Airbyte Engineering <engineering@airbyte.io>