fix: input wrapper registration (#16534)
## What
I had broken the `onBlur` validation handling for `input` type FormControls in previous changes to `InputWrapper`, since it was changed to no longer call `register()`.

## How
The changes in this PR go back to calling `register()` to fix `onBlur` validation mode handling, while also keeping the `useWatch()` call so that updates to parent objects are properly picked up by the Input UI component.

This PR also now sets `valueAsNumber` to `true` when the type is `number`, which lets us get rid of all of the extra `handleChange` logic while still writing actual numeric values to the form as desired.

The downside is that this causes some annoying behavior where if you type in something like `12e4`, then when you press `e`, the input gets cleared out. This is happening because `12e` is not a valid number, so react-hook-form writes `null` to the form values in this case since it can't parse the value into a number, which clears out the UI input since it is being set to the result of `useWatch()`.

I think keeping the useWatch so that this Input is properly updated when a parent object is updated feels more important than properly handling the `e` case, so I'm okay with this solution. But open to suggestions on alternatives if there are any!